requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
PyJWT==2.10.0
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
